Add files via upload
In this project, I implemented a red-eye detection and correction algorithm using OpenCV. The program identifies the red-eye effect in photographs and applies image processing techniques to automatically correct it. This project is part of my image processing module and showcases my ability to work with Python and OpenCV to solve practical image-related problems.